app/requirements.txt
flask==3.0.0
gunicorn==21.2.0
opentelemetry-distro==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-sdk==1.27.0
requests==2.31.0
python-json-logger==2.0.7

